refactor: remove chord step wrapper layer
